classe vaga na verdade é uma classe dao enquanto a classe vaga dao funcionava como dto foram criadas classe dto excluidas e renomeadas para facilitar compreensão
